MFH: Fixed bug #25316 (Possible infinite loop inside _php_stream_write()).

+++ main/streams.c
@@ -20,7 +20,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: streams.c,v 1.125.2.80 2003/09/10 00:59:46 iliaa Exp $ */
+/* $Id: streams.c,v 1.125.2.81 2003/09/11 05:07:47 iliaa Exp $ */
 
 #define _GNU_SOURCE
 #include "php.h"
@@ -908,7 +908,8 @@ PHPAPI size_t _php_stream_write(php_stream *stream, const char *buf, size_t coun
 		} else {
 			justwrote = stream->ops->write(stream, buf, towrite TSRMLS_CC);
 		}
-		if (justwrote > 0) {
+		/* convert justwrote to an integer, since normally it is unsigned */
+		if ((int)justwrote > 0) {
 			buf += justwrote;
 			count -= justwrote;
 			didwrite += justwrote;
